Fix up Epub Javascript object
We had been assigning functions to names without the "var"
specification.  This ends up making them global, and thus methods of the
window object.  They were closures over the local variables, though, so
those references still worked.  We were also passing window to the
callback after the initialization instead of the object we created.
This counteracted our mistake, letting things work.

To attach the methods to the local object, we use "var" for them as
well.  This makes them private, so the methods that Monocle needs are
instead specified with this.name = function ().  However, this makes
them hard to access internally (since this isn't the object in
closures), so we also make private copies for those we need internally.